refactor(gce): move the GoogleServerGroupManagers and related classes
