+++ clouddriver-google/src/main/groovy/com/netflix/spinnaker/clouddriver/google/compute/ZoneGoogleServerGroupManagers.java
@@ -1,4 +1,4 @@
-package com.netflix.spinnaker.clouddriver.google.deploy.instancegroups;
+package com.netflix.spinnaker.clouddriver.google.compute;
 
 import com.google.api.services.compute.Compute;
 import com.google.api.services.compute.ComputeRequest;
