+++ clouddriver-google/src/main/groovy/com/netflix/spinnaker/clouddriver/google/deploy/ops/SetStatefulDiskAtomicOperation.java
@@ -10,8 +10,8 @@ import com.netflix.spinnaker.clouddriver.data.task.Task;
 import com.netflix.spinnaker.clouddriver.data.task.TaskRepository;
 import com.netflix.spinnaker.clouddriver.google.deploy.GCEUtil;
 import com.netflix.spinnaker.clouddriver.google.deploy.description.SetStatefulDiskDescription;
-import com.netflix.spinnaker.clouddriver.google.deploy.instancegroups.GoogleServerGroupManagers;
-import com.netflix.spinnaker.clouddriver.google.deploy.instancegroups.GoogleServerGroupManagersFactory;
+import com.netflix.spinnaker.clouddriver.google.compute.GoogleServerGroupManagers;
+import com.netflix.spinnaker.clouddriver.google.compute.GoogleServerGroupManagersFactory;
 import com.netflix.spinnaker.clouddriver.google.model.GoogleServerGroup;
 import com.netflix.spinnaker.clouddriver.google.provider.view.GoogleClusterProvider;
 import java.io.IOException;
